still map to bfloat

+++ glide_text2im/nn.py
@@ -15,7 +15,7 @@ class GroupNorm32(nn.GroupNorm):
         self.swish = swish
 
     def forward(self, x):
-        y = super().forward(x)
+        y = super().forward(x).to(x.dtype)
         if self.swish == 1.0:
             y = F.silu(y)
         elif self.swish:
